Removed the soft currency converters in Amount.
The problem with these is that they would not work properly on
different currencies which operate on different scale. This also means
that when we log amounts to carbon, we have to append the currency to
the logging key in order to avoid mixing the amounts of two different
currencies.